Update produce.mdx: Clarify meaning of "local mutation"
I didn't relate to the term "local mutation" and that could just be a function of my experience (or lack thereof) but I asked about it on Discord and got a great answer from @fabiospampinato: Local mutation, at lease in this context, means "[you can] mutate the store inside the produce function". 

I'm suggesting changed the doc to remove the term "local mutation" and replace it with @fabiospampinto's definition. However, if you think "local mutation" is relatable to most readers of the docs just let me know and I'll close the issue.

Discord link: https://discord.com/channels/722131463138705510/861229287868858379/1348946097430728725