update Modifying values in arrays / Appending new values simplify.
current 

```Appending new values
To append a new element to an array within a store, you specify the target array and set the index to the desired position. For example, if you wanted to append the new element to the end of the array, you would set the index to array.length:
```

this phrasing may be a little ambiguous, the example for appending, using the path syntax, will always use the array.length 
to append a value. 

> and set the index to the desired position
may imply to readers that read this section without prior context
that the index can be used as insertion point in the array. 
not realizing that only works with non existing index such as the length, or higher index